refactor: StoreRegistration을 ShopRegistration으로 변경
