+++ src/page/ShopRegistration/component/ConfirmPopup/ConfirmPopup.module.scss
