+++ src/page/ShopRegistration/component/InputBox/index.tsx
