+++ src/page/ShopRegistration/component/Modal/Category/Category.module.scss
