+++ src/page/ShopRegistration/component/Modal/ConfirmShop/ConfirmShop.module.scss
