+++ src/page/ShopRegistration/component/Modal/ConfirmShop/index.tsx
