+++ src/page/ShopRegistration/component/Modal/OperateTimeMobile/index.tsx
@@ -1,7 +1,7 @@
 import PreviousStep from 'component/common/Auth/PreviousStep';
 import SubTitle from 'component/common/Auth/SubTitle';
 import useStepStore from 'store/useStepStore';
-import TimePicker from 'page/StoreRegistration/component/TimePicker';
+import TimePicker from 'page/ShopRegistration/component/TimePicker';
 import WEEK from 'utils/constant/week';
 import { createPortal } from 'react-dom';
 import styles from './OperateTimeMobile.module.scss';
