+++ src/page/ShopRegistration/component/Modal/TimeSelection/TimeSelection.module.scss
