+++ src/page/ShopRegistration/component/TimePicker/index.tsx
@@ -5,7 +5,7 @@ import {
 import useModalStore from 'store/modalStore';
 import WEEK from 'utils/constant/week';
 import cn from 'utils/ts/className';
-import TimeSelection from 'page/StoreRegistration/component/Modal/TimeSelection';
+import TimeSelection from 'page/ShopRegistration/component/Modal/TimeSelection';
 import styles from './TimePicker.module.scss';
 
 type Weekday = typeof WEEK[number];
