+++ src/page/ShopRegistration/index.tsx
