+++ src/page/ShopRegistration/view/Mobile/Main/Main.module.scss
