+++ src/page/ShopRegistration/view/Mobile/Main/index.tsx
