+++ src/page/ShopRegistration/view/Mobile/ShopCategory/ShopCategory.module.scss
