+++ src/page/ShopRegistration/view/Mobile/ShopCategory/index.tsx
