+++ src/page/ShopRegistration/view/Mobile/ShopConfirmation/ShopConfirmation.module.scss
