+++ src/page/ShopRegistration/view/Mobile/ShopConfirmation/index.tsx
