+++ src/page/ShopRegistration/view/Mobile/ShopEntry/ShopEntry.module.scss
