+++ src/page/ShopRegistration/view/Mobile/ShopRegistrationMobile.module.scss
