+++ src/page/ShopRegistration/view/Mobile/Sub/Sub.module.scss
