+++ src/page/ShopRegistration/view/PC/ShopRegistrationPC.module.scss
