+++ src/page/ShopRegistration/view/PC/index.tsx
@@ -9,11 +9,11 @@ import useStepStore from 'store/useStepStore';
 import Copyright from 'component/common/Copyright';
 import CustomButton from 'page/Auth/Signup/component/CustomButton';
 import Complete from 'component/common/Auth/Complete';
-import InputBox from 'page/StoreRegistration/component/InputBox';
-import Category from 'page/StoreRegistration/component/Modal/Category';
-import SearchShop from 'page/StoreRegistration/component/Modal/SearchShop';
-import OperateTimePC from 'page/StoreRegistration/component/Modal/OperateTimePC';
-import ConfirmPopup from 'page/StoreRegistration/component/ConfirmPopup';
+import InputBox from 'page/ShopRegistration/component/InputBox';
+import Category from 'page/ShopRegistration/component/Modal/Category';
+import SearchShop from 'page/ShopRegistration/component/Modal/SearchShop';
+import OperateTimePC from 'page/ShopRegistration/component/Modal/OperateTimePC';
+import ConfirmPopup from 'page/ShopRegistration/component/ConfirmPopup';
 import useMediaQuery from 'utils/hooks/useMediaQuery';
 import useBooleanState from 'utils/hooks/useBooleanState';
 import CustomModal from 'component/common/CustomModal';
